--- app/src/main/java/com/example/user/biblejournal/core/Test.java
@@ -1,4 +1,0 @@
-package com.example.user.biblejournal.core;
-
-public class Test {
-}
