+++ app/src/main/java/com/example/user/biblejournal/core/ui/BaseFragment.kt
@@ -19,4 +19,8 @@ open class BaseFragment : Fragment() {
     fun navigate(navId : Int) {
         navController.navigate(navId)
     }
+
+    fun navigate(navId : Int, bundle : Bundle) {
+        navController.navigate(navId, bundle)
+    }
 }
