+++ app/src/main/java/com/example/user/biblejournal/core/util/StringUtil.kt
@@ -1,4 +1,4 @@
-package com.example.user.biblejournal.utils
+package com.example.user.biblejournal.core.util
 
 object StringUtil {
     fun isEmpty(string: String?): Boolean {
