+++ app/src/main/java/com/example/user/biblejournal/editnote/local/entity/NoteEntity.kt
@@ -1,4 +1,4 @@
-package com.example.user.biblejournal.database.note
+package com.example.user.biblejournal.editnote.local.entity
 
 
 import androidx.room.ColumnInfo
