--- app/src/main/java/com/example/user/biblejournal/home/Test.java
@@ -1,4 +1,0 @@
-package com.example.user.biblejournal.home;
-
-public class Test {
-}
